Solved the angle distirtion problem
Issue was that by spacing the rays equally, the points they strike ona  wall are not equally spaced,
bring mostly bunched up near the middle. This problem is more obvious at higher FOVs.
Solution was to compute the angle based on how far we have scanned across the screen, ensuring the
ray collisions are equally spaced on a wall perpindicular to the view point of the player
(while the angles themselves are no longer equally spaced, though they are pretty close.)